Async on sync command. CLI logs moved to log.js
